Changes to matrix_bot_listener.py for new date-time
